+++ Plutarch/BLS.hs
@@ -28,7 +28,8 @@ module Plutarch.BLS (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Bool (PBool, PEq, (#==))
+import Plutarch.Bool (PEq, (#==))
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Lift (DeriveBuiltinPLiftable, PLiftable, PLifted (PLifted), pconstant)