+++ Plutarch/BitString.hs
@@ -13,8 +13,9 @@ module Plutarch.BitString (
 ) where
 
 import GHC.Generics (Generic)
-import Plutarch.Bool (PBool, PEq, POrd, PPartialOrd, pif, (#<))
+import Plutarch.Bool (PEq, POrd, PPartialOrd, pif, (#<))
 import Plutarch.Builtin (PBuiltinList)
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Integer (PInteger)
 import Plutarch.Internal.Newtype (PlutusTypeNewtype)