+++ Plutarch/Crypto.hs
@@ -14,7 +14,7 @@ module Plutarch.Crypto (
   pverifySchnorrSecp256k1Signature,
 ) where
 
-import Plutarch.Bool (PBool)
+import Plutarch.Builtin.Bool (PBool)
 import Plutarch.ByteString (PByteString)
 import Plutarch.Internal.Term (Term, (:-->))
 import Plutarch.Unsafe (punsafeBuiltin)